Reformat the code and fix errors
